requirements.txt
asgiref==3.7.2
Django>=3.0.0
django-widget-tweaks==1.5.0
pytz==2024.1
sqlparse==0.4.4
Pillow==11.2.1
geopy==2.4.1
django-otp>=1.1.0
twilio>=7.0.0
python-dotenv>=0.19.0
django-phonenumber-field>=5.2.0
phonenumbers>=8.12.0

